Extract chip-specific code to modules
